requirements.txt
PySide6>=6.5.0
pyqtgraph>=0.13.3
numpy<1.24.0
requests>=2.31.0
tradingview-ta>=3.3.0
ollama>=0.1.0
PyQt5-sip>=12.11.0
